refactor: move SkipUnmatchedRoutes state into skipRouteIndex
The feature added seven fields to the App struct. Group them into a dedicated
skipRouteIndex struct and move all related logic into a new router_skip.go:

- skipRouteIndex holds the lookahead indexes (staticMethods, bucketParamMethods,
  paramRoutes, hasUseRoutes, hasParamUse) plus the 405-fallback method mask
  (routeMethods, methodMaskValid).
- App now carries a single `skip skipRouteIndex` field instead of seven.
- resolveSkip, emitSkip, emitSkipCustom, buildSkipIndexes, indexedRoute and
  skipResult move to router_skip.go; buildTree just calls app.buildSkipIndexes().

Pure reorganization; no behavior change. Full suite and golangci-lint pass.
