+++ router.go
@@ -240,7 +240,7 @@ func (app *App) next(c *DefaultCtx) (bool, error) {
 	// iteration. They are only mutated outside this loop (before next() is called).
 	firstMatchIndex := c.firstMatchIndex
 	skipNonUse := c.shouldSkipNonUseRoutes
-	skipHasParamUse := app.skipHasParamUse
+	skipHasParamUse := app.skip.hasParamUse
 
 	// Loop over the route stack starting from previous index
 	for indexRoute < lenr {
@@ -316,8 +316,8 @@ func (app *App) next(c *DefaultCtx) (bool, error) {
 
 	exists := false
 	methods := app.config.RequestMethods
-	prune := app.methodMaskValid
-	routeMethods := app.routeMethods
+	prune := app.skip.methodMaskValid
+	routeMethods := app.skip.routeMethods
 	for i := range methods {
 		// Skip original method
 		if methodInt == i {
@@ -385,7 +385,7 @@ func (app *App) nextCustom(c CustomCtx) (bool, error) {
 	values := c.getValues()
 	firstMatchIndex := c.getFirstMatchIndex()
 	skipNonUse := c.getSkipNonUseRoutes()
-	skipHasParamUse := app.skipHasParamUse
+	skipHasParamUse := app.skip.hasParamUse
 
 	// Loop over the route stack starting from previous index
 	for indexRoute < lenr {
@@ -459,8 +459,8 @@ func (app *App) nextCustom(c CustomCtx) (bool, error) {
 
 	exists := false
 	methods := app.config.RequestMethods
-	prune := app.methodMaskValid
-	routeMethods := app.routeMethods
+	prune := app.skip.methodMaskValid
+	routeMethods := app.skip.routeMethods
 	for i := range methods {
 		// Skip original method
 		if methodInt == i {
@@ -509,124 +509,6 @@ func (app *App) nextCustom(c CustomCtx) (bool, error) {
 	return false, ErrNotFound
 }
 
-// SkipUnmatchedRoutes decision codes returned by resolveSkip.
-const (
-	skipRunChain  = iota // proceed to the normal middleware/handler chain
-	skipNotFound         // answer 404 without running the chain
-	skipMethodNot        // answer 405 without running the chain; allowMask holds the methods
-)
-
-// skipResult is the outcome of the SkipUnmatchedRoutes lookahead.
-type skipResult struct {
-	decision   int    // skipRunChain, skipNotFound, or skipMethodNot
-	allowMask  uint64 // methods to advertise in the Allow header (skipMethodNot only)
-	matchIndex int    // pre-resolved endpoint index for next()/nextCustom(), or -1
-}
-
-// resolveSkip implements the SkipUnmatchedRoutes two-tier fast path. matchIndex
-// lets next/nextCustom skip re-checking the endpoints already ruled out. values
-// is used as scratch; on a parametric match it holds that route's params, but
-// next() re-matches the route to recompute them (middleware between here and the
-// endpoint may overwrite the slice). Method masks are uint64 so they stay 64-bit
-// wide on 32-bit platforms.
-func (app *App) resolveSkip(methodInt, treeHash int, detectionPath, path string, values *[maxParams]string) skipResult {
-	methodBit := uint64(1) << methodInt
-	staticMask := app.staticRouteMethods[detectionPath]
-
-	// Tier 1a: a static endpoint matches this method -> run the chain normally.
-	if staticMask&methodBit != 0 {
-		return skipResult{decision: skipRunChain, matchIndex: -1}
-	}
-
-	// Resolve this method's parametric candidates for the request's bucket. paramRoutes
-	// has an entry for every treeStack bucket, so a lookup miss mirrors a treeStack miss
-	// and selects the bucket-0 fallback exactly as next() does; the candidate indices
-	// then line up with the bucket next() iterates.
-	cands, ok := app.paramRoutes[methodInt][treeHash]
-	if !ok {
-		cands = app.paramRoutes[methodInt][0]
-	}
-
-	// Tier 2: scan this method's candidates first, so a match never pays for the
-	// authoritative check below.
-	for _, cand := range cands {
-		if cand.route.match(detectionPath, path, values) {
-			return skipResult{decision: skipRunChain, matchIndex: cand.idx}
-		}
-	}
-
-	// No match for the requested method. If no method has a parametric route reachable
-	// for this bucket (the specific bucket, or bucket 0 for methods that lack it), the
-	// static index is authoritative and no cross-method scan is needed.
-	if app.bucketParamMethods[treeHash]|app.bucketParamMethods[0] == 0 {
-		if staticMask != 0 {
-			return skipResult{decision: skipMethodNot, allowMask: staticMask, matchIndex: -1}
-		}
-		return skipResult{decision: skipNotFound, matchIndex: -1}
-	}
-
-	// Combine the static index with a parametric scan of the other methods to decide
-	// between 405 and 404.
-	allow := staticMask
-	methods := app.config.RequestMethods
-	for m := range methods {
-		if m == methodInt || allow&(uint64(1)<<m) != 0 {
-			continue
-		}
-		c2, ok := app.paramRoutes[m][treeHash]
-		if !ok {
-			c2 = app.paramRoutes[m][0]
-		}
-		for _, cand := range c2 {
-			if cand.route.match(detectionPath, path, values) {
-				allow |= uint64(1) << m
-				break
-			}
-		}
-	}
-	if allow != 0 {
-		return skipResult{decision: skipMethodNot, allowMask: allow, matchIndex: -1}
-	}
-	return skipResult{decision: skipNotFound, matchIndex: -1}
-}
-
-// emitSkip renders a SkipUnmatchedRoutes short-circuit response (404 or 405)
-// through the configured error handler, matching the behavior of next()'s own
-// terminal 404/405 path but without running the middleware chain. It takes the
-// concrete *DefaultCtx so the variadic Append calls stay on the stack (calling
-// Append through the Ctx interface forces the value slice to escape and allocate).
-func (app *App) emitSkip(c *DefaultCtx, allowMask uint64, err error) {
-	// allowMask is only non-zero for the 405 case.
-	if allowMask != 0 {
-		methods := app.config.RequestMethods
-		for i := range methods {
-			if allowMask&(uint64(1)<<i) != 0 {
-				c.Append(HeaderAllow, methods[i])
-			}
-		}
-	}
-	if catch := app.ErrorHandler(c, err); catch != nil {
-		_ = c.SendStatus(StatusInternalServerError) //nolint:errcheck // Always return nil
-	}
-}
-
-// emitSkipCustom is the CustomCtx counterpart of emitSkip for the custom-context
-// request path. The Append calls allocate here (interface dispatch), but this
-// path is rare compared to the default *DefaultCtx handler.
-func (app *App) emitSkipCustom(c CustomCtx, allowMask uint64, err error) {
-	if allowMask != 0 {
-		methods := app.config.RequestMethods
-		for i := range methods {
-			if allowMask&(uint64(1)<<i) != 0 {
-				c.Append(HeaderAllow, methods[i])
-			}
-		}
-	}
-	if catch := app.ErrorHandler(c, err); catch != nil {
-		_ = c.SendStatus(StatusInternalServerError) //nolint:errcheck // Always return nil
-	}
-}
-
 func (app *App) defaultRequestHandler(rctx *fasthttp.RequestCtx) {
 	ctx, ok := app.acquireDefaultCtx(rctx)
 	if !ok {
@@ -651,7 +533,7 @@ func (app *App) defaultRequestHandler(rctx *fasthttp.RequestCtx) {
 	// Skip unmatched routes before running the middleware chain. When no middleware
 	// is registered the lookahead is pure overhead (next() already answers 404/405
 	// without running anything), so it is gated on skipHasUseRoutes.
-	if app.config.SkipUnmatchedRoutes && app.skipHasUseRoutes {
+	if app.config.SkipUnmatchedRoutes && app.skip.hasUseRoutes {
 		res := app.resolveSkip(ctx.methodInt, ctx.treePathHash,
 			utils.UnsafeString(ctx.detectionPath), utils.UnsafeString(ctx.path), &ctx.values)
 		switch res.decision {
@@ -694,7 +576,7 @@ func (app *App) customRequestHandler(rctx *fasthttp.RequestCtx) {
 	// Skip unmatched routes before running the middleware chain. When no middleware
 	// is registered the lookahead is pure overhead (next() already answers 404/405
 	// without running anything), so it is gated on skipHasUseRoutes.
-	if app.config.SkipUnmatchedRoutes && app.skipHasUseRoutes {
+	if app.config.SkipUnmatchedRoutes && app.skip.hasUseRoutes {
 		res := app.resolveSkip(ctx.getMethodInt(), ctx.getTreePathHash(),
 			ctx.getDetectionPath(), ctx.Path(), ctx.getValues())
 		switch res.decision {
@@ -1121,26 +1003,9 @@ func (app *App) buildTree() *App {
 		app.treeStack[method] = tsMap
 	}
 
-	// Precompute the bitmask of methods that own at least one non-use route, so the
-	// 404/405 cross-method fallback in next()/nextCustom() can skip methods that can
-	// never match. Trusted only when the method set fits in a 64-bit mask.
-	app.methodMaskValid = len(app.config.RequestMethods) <= 64
-	app.routeMethods = 0
-	if app.methodMaskValid {
-		for method := range app.config.RequestMethods {
-			for _, route := range app.stack[method] {
-				if !route.use {
-					app.routeMethods |= uint64(1) << method
-					break
-				}
-			}
-		}
-	}
-
-	// Build the SkipUnmatchedRoutes lookup indexes from the freshly built tree.
-	if app.config.SkipUnmatchedRoutes {
-		app.buildSkipIndexes()
-	}
+	// Rebuild the route-resolution indexes (405-fallback method mask and, when
+	// enabled, the SkipUnmatchedRoutes lookahead) from the freshly built tree.
+	app.buildSkipIndexes()
 
 	// reset the flag and return
 	app.hasRoutesRefreshed = false
@@ -1152,90 +1017,4 @@ func reuseRouteBucket(prev map[int][]*Route, key, capHint int) []*Route {
 		return bucket[:0]
 	}
 	return make([]*Route, 0, capHint)
-}
-
-// indexedRoute pairs a candidate route with its index in the tree bucket it was
-// taken from, so the SkipUnmatchedRoutes lookahead can hand that index to
-// next()/nextCustom() (via firstMatchIndex) to skip endpoints already ruled out.
-type indexedRoute struct {
-	route *Route
-	idx   int
-}
-
-// buildSkipIndexes builds the lookup structures used by the SkipUnmatchedRoutes
-// fast path: a method-global index of static endpoints, a per-tree-bucket bitmask
-// of methods that own parametric/root/star endpoints, and a per-bucket candidate
-// list of those endpoints. It is rebuilt together with the tree so the indexes
-// always reflect the current route set.
-func (app *App) buildSkipIndexes() {
-	// The per-method masks are 64-bit. If RequestMethods exceeds 64, method indexes
-	// >= 64 would shift out of range and miscompute, so leave the fast path disabled
-	// (skipHasUseRoutes stays false) and let next() answer 404/405 as usual.
-	if len(app.config.RequestMethods) > 64 {
-		app.staticRouteMethods = nil
-		app.bucketParamMethods = nil
-		app.paramRoutes = nil
-		app.skipHasUseRoutes = false
-		app.skipHasParamUse = false
-		return
-	}
-
-	static := make(map[string]uint64)
-	bucketParam := make(map[int]uint64)
-	paramRoutes := make([]map[int][]indexedRoute, len(app.config.RequestMethods))
-	hasUse := false
-	hasParamUse := false
-
-	for method := range app.config.RequestMethods {
-		bit := uint64(1) << method
-		paramRoutes[method] = make(map[int][]indexedRoute)
-
-		// Static index from the flat stack. A static endpoint is matched by
-		// route.match via a plain exact compare against route.path, so keying by
-		// route.path and looking up by detectionPath is exact.
-		for _, route := range app.stack[method] {
-			if route.use {
-				hasUse = true
-				// A parametric or wildcard middleware writes into c.values when it
-				// matches, which would clobber the lookahead's params.
-				if route.star || len(route.Params) > 0 {
-					hasParamUse = true
-				}
-				continue
-			}
-			if route.mount || route.star || route.root || len(route.Params) > 0 {
-				continue
-			}
-			static[route.path] |= bit
-		}
-
-		// Per-bucket candidate list of parametric/root/star endpoints from the
-		// final buckets (post bucket-0 replication) so they mirror exactly what
-		// next() scans per bucket. The stored idx is the route's position in that
-		// same bucket, which next() iterates. An entry is created for every bucket
-		// (empty when it has no candidates) so a lookup miss in resolveSkip mirrors
-		// a treeStack miss and selects the bucket-0 fallback. bucketParamMethods
-		// records which methods have candidates per bucket, for the authoritative
-		// no-scan miss. Mounted sub-apps are expanded into normal routes before
-		// buildTree runs, so no mount route reaches this point.
-		for treeHash, bucket := range app.treeStack[method] {
-			var cands []indexedRoute
-			for i, route := range bucket {
-				if route.use || route.mount {
-					continue
-				}
-				if route.root || route.star || len(route.Params) > 0 {
-					cands = append(cands, indexedRoute{route: route, idx: i})
-					bucketParam[treeHash] |= bit
-				}
-			}
-			paramRoutes[method][treeHash] = cands
-		}
-	}
-
-	app.staticRouteMethods = static
-	app.bucketParamMethods = bucketParam
-	app.paramRoutes = paramRoutes
-	app.skipHasUseRoutes = hasUse
-	app.skipHasParamUse = hasParamUse
 }
